refactor: merge to get-artifacts
