+++ lib/utils/get-artifacts.ts
@@ -1,5 +1,46 @@
+import fs from 'fs/promises';
+import path from 'path';
+import { readPackageUpAsync } from 'read-pkg-up';
+import type { Artifact } from '../types';
 import artifactPaths from '../artifacts';
-import { getArtifact } from './get-artifact';
+import { getSize, getGzipSize } from './get-size';
+
+const readPublicPackageUp = async (cwd: string) => {
+	let packageFound = await readPackageUpAsync({ cwd });
+	while (packageFound.packageJson.private) {
+		packageFound = await readPackageUpAsync({
+			cwd: path.join(path.dirname(packageFound.path), '..'),
+		});
+	}
+	return packageFound.packageJson;
+};
+
+const isFilePathPattern = /^[./]/;
+
+export const getArtifact = async (
+	modulePath: string,
+): Promise<Artifact> => {
+	if (!isFilePathPattern.test(modulePath)) {
+		modulePath = require.resolve(modulePath);
+	}
+
+	const [
+		packageJson,
+		code,
+	] = await Promise.all([
+		readPublicPackageUp(modulePath),
+		fs.readFile(modulePath),
+	]);
+
+	return {
+		moduleName: packageJson.name,
+		moduleVersion: packageJson.version,
+		modulePath,
+		code,
+		size: getSize(code),
+		gzipSize: getGzipSize(code),
+	};
+};
 
 export const getArtifacts = async () => {
 	const artifacts = await Promise.all(artifactPaths.map(
